# docs/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Candidate Compliance Platform API",
    "version": "1.0.0",
    "description": "Product-neutral REST API for tenant-scoped candidate compliance, verification, and governed CV extraction. Tenant identifiers supplied by clients are validated against the authenticated actor's memberships before domain logic runs."
  },
  "servers": [
    { "url": "http://localhost:4000", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/v1/candidates/{candidateId}/documents": {
      "parameters": [{ "$ref": "#/components/parameters/CandidateId" }],
      "post": {
        "operationId": "createComplianceDocument",
        "summary": "Create a compliance document",
        "description": "Creates a logical document and its first immutable-history version as DRAFT.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:create",
        "parameters": [
          { "$ref": "#/components/parameters/TenantIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceDocumentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compliance document created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComplianceDocument" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "get": {
        "operationId": "listCandidateComplianceDocuments",
        "summary": "List a candidate's compliance documents",
        "description": "Lists active logical documents and their current versions for one active candidate.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:read",
        "parameters": [
          { "$ref": "#/components/parameters/TenantIdHeader" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "$ref": "#/components/parameters/DocumentTypeFilter" },
          { "$ref": "#/components/parameters/DocumentStatusFilter" }
        ],
        "responses": {
          "200": {
            "description": "Paginated compliance documents.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceDocumentListResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/documents/expiring": {
      "get": {
        "operationId": "listExpiringComplianceDocuments",
        "summary": "List documents expiring within 30 days",
        "description": "Returns active logical documents whose current version expires from the current UTC date through day 30, inclusive.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:read",
        "parameters": [
          { "$ref": "#/components/parameters/TenantIdHeader" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "$ref": "#/components/parameters/DocumentTypeFilter" },
          { "$ref": "#/components/parameters/DocumentStatusFilter" }
        ],
        "responses": {
          "200": {
            "description": "Paginated expiring compliance documents.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceDocumentListResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/documents/{documentId}": {
      "parameters": [{ "$ref": "#/components/parameters/DocumentId" }],
      "get": {
        "operationId": "getComplianceDocument",
        "summary": "Get a compliance document",
        "description": "Returns an active logical document and its current version.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:read",
        "parameters": [{ "$ref": "#/components/parameters/TenantIdHeader" }],
        "responses": {
          "200": {
            "description": "Compliance document found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComplianceDocument" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "operationId": "removeComplianceDocument",
        "summary": "Remove a compliance document logically",
        "description": "Performs one-way retention-safe logical removal of the stable aggregate. Versions, supersession links, verification history, and audit evidence remain retained and unchanged. Exact same-key replay returns 204 without another mutation.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:remove",
        "parameters": [
          { "$ref": "#/components/parameters/TenantIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" }
        ],
        "responses": {
          "204": {
            "description": "Document removed logically. No response body."
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/documents/{documentId}/versions": {
      "parameters": [{ "$ref": "#/components/parameters/DocumentId" }],
      "get": {
        "operationId": "listComplianceDocumentVersions",
        "summary": "List document version history",
        "description": "Returns the complete persisted version history in ascending version-number order. Exactly one item identifies the current version. Internal tenant, creator and supersession identifiers are not exposed.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:read",
        "parameters": [{ "$ref": "#/components/parameters/TenantIdHeader" }],
        "responses": {
          "200": {
            "description": "Complete immutable version history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceDocumentVersionHistoryResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "operationId": "createComplianceDocumentVersion",
        "summary": "Create a document version",
        "description": "Appends a DRAFT version and makes it current. This general operation cannot overwrite or supersede an approved current version; use correction instead.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:create",
        "parameters": [
          { "$ref": "#/components/parameters/TenantIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceDocumentVersionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Version created and selected as current.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComplianceDocument" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/documents/{documentId}/approve": {
      "parameters": [{ "$ref": "#/components/parameters/DocumentId" }],
      "post": {
        "operationId": "approveComplianceDocument",
        "summary": "Approve the current document version",
        "description": "Transitions a DRAFT or PENDING_REVIEW current version to APPROVED. Exact same-key replay returns the stored response without duplicate audit. A new key targeting an already-approved current version succeeds as a semantic no-op, appends immutable approval evidence with equal before/after hashes and ALREADY_APPROVED metadata, and does not mutate document state.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:approve",
        "parameters": [
          { "$ref": "#/components/parameters/TenantIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmptyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current version approved or already approved.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComplianceDocument" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/documents/{documentId}/corrections": {
      "parameters": [{ "$ref": "#/components/parameters/DocumentId" }],
      "post": {
        "operationId": "correctComplianceDocument",
        "summary": "Correct an approved document version",
        "description": "Creates a new DRAFT that supersedes the approved current version and becomes current. The approved version is retained unchanged as immutable history.",
        "tags": ["Compliance documents"],
        "security": [{ "bearerAuth": [] }],
        "x-required-permission": "document:correct",
        "parameters": [
          { "$ref": "#/components/parameters/TenantIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceDocumentCorrectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Corrected DRAFT version created and selected as current.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ComplianceDocument" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT returned by POST /api/v1/auth/login."
      }
    },
    "parameters": {
      "TenantIdHeader": {
        "name": "X-Tenant-Id",
        "in": "header",
        "required": true,
        "description": "Untrusted tenant selection validated against the authenticated actor's current memberships.",
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
      "IdempotencyKeyHeader": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Opaque key scoped by validated tenant membership and operation. Same input replays the stored response; conflicting reuse returns 409.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 200,
          "pattern": "^[A-Za-z0-9._~:/+\\-]+$",
          "example": "candidate-create-20260815-001"
        }
      },
      "CandidateId": {
        "name": "candidateId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
      "DocumentId": {
        "name": "documentId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/Uuid" }
      },
...
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100000,
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
      "DocumentTypeFilter": {
        "name": "type",
        "in": "query",
        "required": false,
        "schema": { "$ref": "#/components/schemas/ComplianceDocumentType" }
      },
      "DocumentStatusFilter": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": { "$ref": "#/components/schemas/ComplianceDocumentStatus" }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Request validation, tenant selection, upload, or idempotency key failed.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "Unauthorized": {
        "description": "Credentials or Bearer authentication are invalid.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "Forbidden": {
        "description": "Tenant context is unavailable or the operation-specific permission is denied.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "NotFound": {
        "description": "The resource is missing, inactive, or unavailable in the selected tenant.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
      "Conflict": {
        "description": "A domain invariant or idempotency-key reuse conflicts with this request.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Unexpected failure without internal details.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      }
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
...
      "PaginationMetadata": {
        "type": "object",
        "additionalProperties": false,
        "required": ["page", "pageSize", "totalItems", "totalPages"],
        "properties": {
          "page": { "type": "integer", "minimum": 1 },
          "pageSize": { "type": "integer", "minimum": 1 },
          "totalItems": { "type": "integer", "minimum": 0 },
          "totalPages": { "type": "integer", "minimum": 0 }
        }
      },
...
      "ComplianceDocumentType": {
        "type": "string",
        "enum": [
          "RIGHT_TO_WORK",
          "BACKGROUND_CHECK",
          "PROFESSIONAL_CERTIFICATION",
          "OTHER"
        ]
      },
      "ComplianceDocumentStatus": {
        "type": "string",
        "enum": ["DRAFT", "PENDING_REVIEW", "APPROVED", "REJECTED"]
      },
      "NullableDate": { "type": ["string", "null"], "format": "date" },
      "ComplianceDocumentCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type"],
        "properties": {
          "type": { "$ref": "#/components/schemas/ComplianceDocumentType" },
          "issueDate": { "$ref": "#/components/schemas/NullableDate" },
          "expiryDate": { "$ref": "#/components/schemas/NullableDate" }
        },
        "description": "When both dates are non-null, expiryDate must not be earlier than issueDate."
      },
      "ComplianceDocumentVersionCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "issueDate": { "$ref": "#/components/schemas/NullableDate" },
          "expiryDate": { "$ref": "#/components/schemas/NullableDate" }
        },
        "description": "Both fields are optional. When both are non-null, expiryDate must not be earlier than issueDate."
      },
      "ComplianceDocumentCorrectionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["issueDate", "expiryDate"],
        "properties": {
          "issueDate": { "$ref": "#/components/schemas/NullableDate" },
          "expiryDate": { "$ref": "#/components/schemas/NullableDate" }
        },
        "description": "Complete corrected date state. When both dates are non-null, expiryDate must not be earlier than issueDate."
      },
      "ComplianceDocumentVersion": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "versionNumber",
          "issueDate",
          "expiryDate",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "versionNumber": { "type": "integer", "minimum": 1 },
          "issueDate": { "$ref": "#/components/schemas/NullableDate" },
          "expiryDate": { "$ref": "#/components/schemas/NullableDate" },
          "status": { "$ref": "#/components/schemas/ComplianceDocumentStatus" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ComplianceDocumentVersionHistoryItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "versionNumber",
          "issueDate",
          "expiryDate",
          "status",
          "createdAt",
          "isCurrent"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "versionNumber": { "type": "integer", "minimum": 1 },
          "issueDate": { "$ref": "#/components/schemas/NullableDate" },
          "expiryDate": { "$ref": "#/components/schemas/NullableDate" },
          "status": { "$ref": "#/components/schemas/ComplianceDocumentStatus" },
          "createdAt": { "type": "string", "format": "date-time" },
          "isCurrent": { "type": "boolean" }
        }
      },
      "ComplianceDocumentVersionHistoryResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "minItems": 1,
            "contains": {
              "type": "object",
              "required": ["isCurrent"],
              "properties": { "isCurrent": { "const": true } }
            },
            "minContains": 1,
            "maxContains": 1,
            "items": {
              "$ref": "#/components/schemas/ComplianceDocumentVersionHistoryItem"
            }
          }
        }
      },
      "ComplianceDocument": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "candidateId",
          "type",
          "currentVersion",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "candidateId": { "$ref": "#/components/schemas/Uuid" },
          "type": { "$ref": "#/components/schemas/ComplianceDocumentType" },
          "currentVersion": {
            "$ref": "#/components/schemas/ComplianceDocumentVersion"
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ComplianceDocumentListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "pagination"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ComplianceDocument" }
          },
          "pagination": { "$ref": "#/components/schemas/PaginationMetadata" }
        }
      },
      "EmptyRequest": {
        "type": "object",
        "additionalProperties": false,
        "maxProperties": 0
      },
...
      "ProblemDetailsError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["path", "message"],
        "properties": {
          "path": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "example": "about:blank"
          },
          "title": { "type": "string", "example": "Bad Request" },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599,
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "The request data is invalid."
          },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProblemDetailsError" }
          }
        }
      }
    }
  }
}